requirements.txt
pynamodb==5.3.0
pytest==7.2.0
pytest-xdist==3.0.2
Faker==15.1.3
factory-boy==3.2.1
